NAS-138691 / 25.10.2 / Add more handling for transient ENOENT on cred check (by anodos325) (#17718)
During health checks for machine account password, the third-party
python-gssapi library may fail to create a ccache object after
successful kinit. In this situation a CallError with errno set to ENOENT
will be raised. This commit converts the error to be non-fatal since by
this point we've validated that the credential is correct. This prevents
the activedirectory service from becoming FAULTED for spurious reasons.

Original PR: https://github.com/truenas/middleware/pull/17716

Co-authored-by: Andrew Walker <andrew.walker@truenas.com>