NAS-137899 / 26.04 / Fix edge case where ix-apps dataset can have invalid mountpoint (#17695)
This commit fixes an issue where if a pool was exported and then
imported, we set the mountpoint for ix-apps ds to where we expect
ix-apps to be mounted but if apps are already configured with a
different pool - this obviously causes problems. This step is not
required here because when we start/configure apps - we ensure that
ix-apps dataset has correct mountpoint set (for the actual pool which is
really configured to be used for apps).