NAS-138569 / 26.04 / NVIDIA support for LXC containers (#17691)
## Context

It was requested to add nvidia support for lxc containers which has been
added. However lxc containers for nvidia also require nvidia drivers to
be available on the host, earlier this implementation lived in docker
plugin which has now been changed and moved to system advanced service
so it can be configured from there and used for both containers and
docker plugin.