NAS-138692 / 26.04 / Fix ipv6 interface autoconf CI check (#17719)
Moving from `dhclient` to `dhcpcd` introduced a change in that the
`sysctl` `autoconf` setting is updated to reflect the current status of
the interface.

By default the IPv6 `autoconf` setting is `1` (enabled). The `dhcpcd`
daemon dynamically updates this based on what routing capability it
senses from the interface. We don't currently route IPv6 in our CI
environment, so the `IPv6` `autoconf` setting is disabled, i.e. set to
`0`.

This PR updates the test to be reactive to the current `IPv6` routing
capability.
This has been tested in a CI run.

NOTE: The VMs under test start as DHCP and get configured with static IP
in `test_005_interface.test_002_configure_interface.`