NAS-138709 / 26.04 / Add SED details in disk.details endpoint (#17725)
This commit adds changes so that we also get sed details in disk.details
endpoint. Basically this endpoint is being used by UI on zpool creation
screens for showing disks. In the current changes, we get SED info from
DB and sed status (if it's SED disk) in realtime from disk.query
endpoint as that already currently efficiently handles this.