Fix restore of SMB share ACLs from config file
This commit changes middleware behavior so that it flushes stored
copies of SMB share ACLs to the share_info.tdb file if the file
is empty. This handles a potential edge case where user is restoring
the TrueNAS config file on a fresh system dataset (for example, fresh
VM install of TrueNAS with data pool imported).