NAS-138429 / 26.04 / Send new UI port to TNC when changed after registration (#17735)
Whenever the UI port is changed via `system.general.update`, a flag will
be set which will be checked whenever the HTTP service is restarted or
reloaded. If set, indicating that the restart will change the effective
UI port, send a PUT request containing the new port to TNC. Retry up to
two times if the first request fails.