NAS-138309 / 26.04 / Make sure containers work in HA (#17727)
## Context

Validation has been added to prevent containers being used in HA if the
system is not licensed for it (license logic has been copied over from
what we had in virt plugin for containers). Also we will try to
gracefully stop containers on failover (like we do for VMs/apps) but
won't block failover and similarly on master we will make sure that if
any container is set to start on boot - it gets started appropriately.