NAS-138766 / 26.04 / Ignore netdata EACCES and EPERM for read_gpt (#17743)
Under certain conditions, TrueNAS can get into a state where `netdata`
spams the syslog a permission failure in `read_gpt`.
This is probably due to the `sd*` devices being owned by `[root,disk]`
with `660` permissions and `netdata` run as the netdata user.

The conditions to reproduce this is not entirely clear. I was able to
reproduce it on a local hyper-v TrueNAS vm, possibly under a
fresh-install condition. I was not able to reproduce it on a Jenkins
generated VM, but I was also not able to re-create the same conditions.

It's possible this is limited to hyper-v VMs, but under the right
conditions it to be reproduced on KVM vms or even real hardware.