Add check to avoid adding duplicate apps entries in acl
This commit adds an additional check when generating an ACL for
datasets with the `APPS` purpose. This dataset purpose minimally
sets permissions such that the apps user (568) has read / write
access to the dataset. In the case where an NFSv4 dataset is
created under an existing NFSv4 dataset and the parent has an
inheritable ACL, middleware calculates the inherited ACL applies
it to the newly created dataset. The original logic was to add
a separate apps entry to the generated ACL before setting it.
This commit changes the logic slightly in that if the dataset
already has an ACL entry that (1) grants the modify permissions
set and (2) is inheritable on new files and directories to the
apps user, the new apps-specific entry will *not* be added.