NAS-138837 / 26.04 / Fix mountpoint issue for containers on importing pools (#17776)
## Problem

When importing a pool, if the mountpoint of container dataset is already
correct - we were falling through `inherit` block where system will
actually change it by adding it to `to_inherit` list.

## Solution

In the event when container dataset is already correct when a pool is
imported, we do not want to do anything in that case.