NAS-138835 / 26.04 / Fix container attachment delegate (#17774)
## Problem

`container.stop` is a job now whereas in the attachment delegate we are
not treating it as one which can mean that when we want to have it
stopped, we will not wait for it to stop which can cause further issues
during the attachment delegate execution flow.

## Solution

Make sure `container.stop` gets treated as a job and is properly stopped
in attachment delegate when called.