NAS-138373 / 26.04 / Return NIC attach choices grouped by type (BRIDGE/MACVLAN) (#17775)
## Context

It was requested by UI team that we should be more descriptive with NIC
attach choices for VM and container devices. Previously, the
`nic_attach_choices` endpoint returned a flat dictionary of all
available interfaces, leaving it to the UI to determine which interfaces
are bridges vs which can be used as MACVLAN parents.

  Example Response
```
  {
    "BRIDGE": ["br0", "br1"],
    "MACVLAN": ["enp0s3", "eth0", "eth1"]
  }
  ```