NAS-138678 / 25.10.2 / Only account for `ONLINE` vdevs when replacing disk (#17728)
We should only account for ONLINE vdevs.

For example, a vdev might be UNAVAIL when its path is /dev/sdb2, but the
current system’s sdb disk may have been replaced with an unrelated one
(with a different partition GUID). We should not include this disk
partition’s size when calculating the new disk partition size.