NAS-138849 / 26.04 / Fix spaces in emails (#17788)
When converting from plain message text to HTML in the mail service, new
lines are replaced with "
", which is sensible.
But whoever did that forgot that in HTML consecutive spaces are counted
as 1 space, causing hugely incorrect and unreadable emails whenever
multiple spaces are involved (ie. basically all console outputs).