Write group SIDs to SMB auditing configuration
This commit converts how we write the SMB auditing configuration
so that the watched or ignored groups will be written to the
configuration by their SID value rather than their names. The
practical impact of this is that local non-SMB groups will not
be allowed for auditing filters. If a problematic auditing
configuration is detected at smb.conf generation then the affected
SMB share(s) will be automatically disabled and an alert generated.