NAS-138854 / 26.04 / convert pool.snapshot.query to use zfs.resource.snapshot.query_impl (#17791)
This converts, arguably, our most inefficient/performance nightmare
endpoint to use the new C module for interacting with snapshots.

For example, on a system that I've been testing this one it has ~8.1k
snapshots.

The legacy endpoint JUST BY CALLING IT:
1. spun a worker process in the pool at 100%
2. the worker process grew > 100MB of resident memory usage
3. took ~64 seconds to complete
4. for reasons that I truly can't be asked to figure out, caused ~60MB
resident memory usage to increase in the parent process that doesn't
seem to immediately be released since it appeared AFTER the worker
process completed

The new endpoint JUST BY CALLING IT:
1. uses the main thread (it spun the parent thread at ~20%)
2. took 0.29 seconds
3. didn't cause resident memory to grow more than even 1KB

To be fair, however, the numbers above are a bit unfair since the legacy
endpoint RECURSIVELY queries snapshots and EVERY SINGLE PROPERTY
AVAILABLE BY DEFAULT. That is truly insane and is what causes the vast
majority of the slowdown. This becomes even more insane when the main
consumer of this endpoint (UI) only uses clones, used, reference and
creation properties and gets the holds information. So running a more
fair test followed by querying all 8.1k snapshots and only retrieving
the 4x properties that the UI team relies on takes around the same time
compared between the 2 endpoints (4-6 seconds). However, the new module
does NOT use ~100MB+ in a child process but instead uses a thread.