NAS-138877 / 26.04 / Properly handle mounting of encrypted dataset children (#17813)
This commit adds handling for mounting the children of encrypted
datasets. The original version of this function from before 26.04 would
pass a recursive argument to the mount operation for each dataset
unlocked. This doesn't make a ton of sense here because we want to be
able to track what is successfully unlocked / mounted verses what errors
occurred and provide results to API consumer.

When the function was refactored to use truenas_pylibzfs, the recursion
was removed. This re-adds some of the recursion for mounts, but only if
dataset shares encryption root and does not have a legacy mountpoint.
While I was in here I made a few minor fixes:

1. replaced shutil.move with os.rename call. We should attempt atomic
operations if possible and otherwise error out.

2. catch cases where dataset is perhaps already mounted (race on shell /
middleware operations was in theory possible).

3. include all the datasets that we've unlocked / mounted in the API
response

Further work / redesign of what we pass to
pool.dataset.unlock_handle_attachments is required but out of scope
here. We should be passing a list of datasets that were actually
unlocked / mounted rather than what API consumer requested.

This unfortunately leads to larger design issues where we are not
storing the dataset information for dataset attachments (NAS-137337).