Add an API endpoint to resync kerberos keytab
This commit adds a new API endpoint `directoryservices.sync_keytab`
that can be used to re-sync the local kerberos keytab with entries
from the remote computer account. The prospective use-case for this
is to allow AD administrators to join TrueNAS to AD, then add
any required SPN entries to the newly-minted AD computer object,
then sync the local truenas keytab up with the AD changes.