NAS-138993 / 26.04 / Do not call `@cache`d function twice (#17854)
Sometimes in the middleware startup logs we see the following line
repeated:
```
[2025/12/18 04:20:21] (INFO) RDMAService._get_link_choices():72 - Fetching RDMA link netdev choices
[2025/12/18 04:20:21] (INFO) RDMAService._get_link_choices():72 - Fetching RDMA link netdev choices
```

`_get_link_choices` is a function wrapped with `@cache`, but if we call
it twice at the same time, data race can occur and the underlying
function will be called twice, which is exactly the thing we want to
prevent. Add the [Double-checked
locking](https://en.wikipedia.org/wiki/Double-checked_locking) to
mitigate that.