NAS-138984 / 26.04 / Enable gssproxy kernel (#17847)
To enable NFS mounts with kerberos security we need the kernel to be
configured for `gssproxy`.
For TrueNAS, kerberos security is valid only when a directory service is
configured, specifically Active Directory or FreeIPA.

Starting the systemd service `auth-rpcgss-module` configures the kernel
to use `gssproxy`.
This configuration can be confirmed by reading
`/proc/net/rpc/use-gss-proxy`
If the file is missing or reports `-1` or `0`, the kernel _is not_
configured for `gssproxy`.
If the file reports `1`, the kernel _is_ configured for `gssproxy`.

The test_directoryservices_basic.py CI test includes a check of this
proc file.

This has been manually tested with two TrueNAS servers in an AD realm
with one acting as the client and the other the server.

---------

Co-authored-by: Caleb St. John <30729806+yocalebo@users.noreply.github.com>
(cherry picked from commit 26e299b40aa096ea9945c42ea475234ebcf81642)