NAS-138984 / 25.10.2 / Enable gssproxy kernel (by mgrimesix) (#17860)
Automatic cherry-pick failed. Please resolve conflicts by running:

    git reset --hard HEAD~1
    git cherry-pick -x 41a8712f7f654546e1c06c5efecbb8762d64066e
    git cherry-pick -x c547b8e283103c0d1b76d6d8c9f28fb533fe2ed3
    git cherry-pick -x 178f117e05948cbbe321aa851154dfd58d5e9cc0
    git cherry-pick -x 35ab9bad5871db418950ac466720657b15744ffb

If the original PR was merged via a squash, you can just cherry-pick the
squashed commit:

    git reset --hard HEAD~1
    git cherry-pick -x 26e299b40aa096ea9945c42ea475234ebcf81642

To enable NFS mounts with kerberos security we need the kernel to be
configured for `gssproxy`.
For TrueNAS, kerberos security is valid only when a directory service is
configured, specifically Active Directory or FreeIPA.

Starting the systemd service `auth-rpcgss-module` configures the kernel
to use `gssproxy`.
This configuration can be confirmed by reading
`/proc/net/rpc/use-gss-proxy`
If the file is missing or reports `-1` or `0`, the kernel _is not_
configured for `gssproxy`.
If the file reports `1`, the kernel _is_ configured for `gssproxy`.

The test_directoryservices_basic.py CI test includes a check of this
proc file.

This has been manually tested with two TrueNAS servers in an AD realm
with one acting as the client and the other the server.

Original PR: https://github.com/truenas/middleware/pull/17847

---------

Co-authored-by: Mark Grimes <126630154+mgrimesix@users.noreply.github.com>
Co-authored-by: Caleb St. John <30729806+yocalebo@users.noreply.github.com>
Co-authored-by: Mark Grimes <mark.grimes@ixsystems.com>