NAS-138985 / 26.04 / Allow importing an offline pool which already is in the database (#17848)
This PR adds changes to introduce a new endpoint which allows
reimporting a zfs pool which exists in the database but hasn't been
imported during boot stage on ZFS side. The use case currently is to
allow reimporting SED based pools which for some reason might not have
unlocked at boot (that should not be the case ideally speaking) but the
same endpoint could also be used to import pools for which for some
reason maybe the disks were not there or something similar and at boot
time the import had failed. Right now the only way to make this happen
would be remove that pool first and then import pool which will then
create a new entry in the database.