NAS-138809 / 26.04 / Properly failover on shutdown in HA (#17833)
## Context

Currently if active node in a HA system is being shutdown, we do not
gracefully handle it and treat it like we do for non-HA systems. This is
problematic because it means that some services can take some time to
shutdown which can potentially mean the other node not coming back
online in time or even pools not importing there because fenced was
still running on the old active node.

After much discussion, we decided that on HA systems priority is uptime.
Which means that if a node is to be rebooted (for system update etc), we
should ensure that the other node comes online as active node as soon as
possible and the best way to make that happen is to mimic what become
passive implementation does. However this is done separately with no
dependency on middleware in the event middleware is down for some
reason.

In short, if a system is HA and failover is not disabled on the system -
on shutdown/reboot of the node, we will panic the system so failover
gets triggered ASAP and the other node comes up as active node.