NAS-139016 / 26.04 / replace zfs.dataset.unlocked_zvols_fast w/ zfs.resource.unlocked_zvols_fast (#17869)
The entire point of the method "unlocked_zvols_fast" was to by pass the
legacy py-libzfs library. However, we added the internal API endpoint
into the `zfs.dataset` namespace. While we're not using the py-libzfs
library, we are, however, creating a child process each time that's
called which is inefficient and almost just as painful as using
py-libzfs.

This adds a new drop-in replacement under `zfs.resource` that does not
use the process pool. It behaves the same way and there should be no
change in behavior and the test passes.