Add API plumbing for stateful failover
This commit adds basic plumbing in the SMB configuration and
validation for stateful SMB failover. This shifts samba's relevant
state tdb files to the .system/samba4 dataset from tmpfs
filesystem. This change is preparatory step for maintaining
server state across failover. The feature as designed is
incompatible with the following features:

1. SMB1 server
2. LEGACY_SHARE and MULTIPROTOCOL_SHARE features
3. Non-HA

The purpose for (2) above is that we want file locking and handles
to be entirely managed by samba itself (for example, no POSIX
locking).

There will be multiple follow-on tickets to this commit in
order to refine the failover process and begin adding SMB server
features simplify session recovery.