NAS-138705 / 26.04 / Add dimensions to truenas pool charts in case of new pools (#17904)
## Problem

The **Netdata pool usage plugin** is unable to update chart
**dimensions** when a new pool is created.
As a result, statistics for newly created pools are not reported, and
the UI cannot display their usage metrics.

## Solution

Update the Netdata pool usage plugin to **dynamically add chart
dimensions** whenever a new pool is detected.
This ensures that usage statistics for newly created pools are collected
correctly and become immediately visible in the UI.