NAS-139076 / 26.04 / Fix parsing / flush of legacy share acl format (#17913)
Between TrueNAS 13 and SCALE we altered the format for SMB share ACLs
that is written to our sqlite3 database from a basic security descriptor
string as formatted by the sharesec utility to the actual secuirty
descriptor bytes that are stored in the share_info.tdb file. This commit
fixes handling for cases where user has the old format written to the DB
and adds tests.