NAS-139114 / 26.04 / fix memory leak in ssh login failure alert (#17923)
Again, analyzing the memory allocation graph of middlewared, I saw
~13.3MB of memory allocated to the `journal.Reader()` object ALONE.
Also, this is a resource leak because the `Reader()` object MUST BE
CLOSED after use which this alert was not doing.

This is particularly painful in this alert because we're creating a new
instance every 60 seconds....painful. Either way, we don't need to use
`journal.Reader()` object for this info. We can get the same information
from our auditd integration. NOTE: this is dependent on
https://github.com/truenas/audit_rules/pull/23.

Finally, there is 0 reason to run this every 60 seconds since we only
care about logs in past 24 hours.