NAS-139127 / 26.04 / remove unused python3-dbus (#17928)
Don't get the warm and fuzzies from this library even if we were using
it... this is from their docs

>dbus-python might not be the best D-Bus binding for you to use.
dbus-python does not follow the principle of “In the face of ambiguity,
refuse the temptation to guess”, and can’t be changed to not do so
without seriously breaking compatibility.
In addition, it uses libdbus (which has known problems with
multi-threaded use) and attempts to be main-loop-agnostic (which means
you have to select a suitable main loop for your application).