NAS-139129 / 26.04 / Fix large memory allocations from 3rd party library (pystemd) (#17929)
I'm analyzing ALL THE MEMORY for middlewared and I found a maddening
chunk of allocations related to pystemd. Investigating shows that every
time we call `unit = self._get_systemd_unit`, it eventually ends up in
`lxml` land. That library causes nothing but grief and I've gone through
and removed it from our product for reasons that also apply here. (I
removed it years ago when we were still on freeBSD for same reasons).

This brings in, yet another, 3rd party dependency but this one is much
cleaner and can be reasoned about. It is pure python and has async
native methods for interacting with dbus.

This implements the necessary methods in pure async fashion so we can
remove the use of `pystemd.systemd1.Unit`.