NAS-139126 / 26.04 / optimize zfs/mount_events.py (#17927)
Still analyzing memory profile of middlewared and this one stuck out as
allocating a rather large amount of memory overtime. This makes sense
because:
1. we're reading entirety of mountinfo twice
2. we're then creating a new list of mountinfo via splitlines()

As an aside, there is no reason to use `os.seek` on the fd. Just use
`f.seek(0)`. It accomplishes the same.

This fixes all of these issues by iterating over the file line by line.
(not loading everything into memory) and only populating the 2x
dictionaries with zfs entries that we care about (instead of
everything).