NAS-139134 / 26.04 / remove pystemd (#17932)
Remove our dependency on `python3-pystemd`. It has too many landmines
that we're running into. It also has a bunch of transitive dependencies
that we don't want. `jeepney` is pure python with no transitive
dependencies. More importantly, the code-base is readable and simple
(and pure python, did I mention that? :smiling)

I was able to test these changes using a standalone script and verify
that we're not leaking memory nor are we leaking fd's.

The benefit is that this is a proper asyncio native implementation for
interacting with dbus and therefore interacting with services.

The only blocking method left in here is the `journal.Reader` class, and
I have an async native solution for that as well but it'll have to come
in another PR since it requires a bit more change.